Fix mac bug where luascript.luac wasn't building
SContruct was modified to run both builders, one for lua and another for
xsdb. The modification was needed because in OS X we link the lua code
with a -setcreate flag, while on linux it is just another .o that ld
links. By seperating both builders we are able to specify when to build
the .o from the luac (for linux) and when to simply leave the .o
parameter out.

Signed-off-by: Ardavon Falls <afalls@mobitv.com>